refactor states; add state manager injection
This is a big change in which most of the state code was changed in one go. Some of the rework is cleanup for readability; mostly putting relevant lines near one another.

This also adds StateManager injection to the various State constructors. It became necessary to give each State object an idea of the broader StateManager.